tests: rename background.py to device_handler.py
